Missed some references to OpenAI in the config module
Made a database.json config for db-migrate since we are no longer
using the DATABASE_URL env var.

Re-made the initial migration and updated the command that's run to
create migrations to add the required option to make the SQL template
files.

Tested the initial setup script and it works.