Added Express skeleton app
Some cleanup of the Prisma schema and adding missing column/table
mappings.

Renamed script that makes data, tested it and it properly write to the
DB. Updated the script to make quarterly periods when the base period
is Quarterly.

Added option to run data gen script to the setup helper script.